IIIF #82 - Replacing "cantaloupe::api" service with "iiif::server" service
